+++ demo/src/Components/UIList/Item.jsx
@@ -3,7 +3,7 @@ import T from 'prop-types'
 import classNames from 'classnames'
 
 const Item = ({ children, className }) => (
-  <li className={classNames('o-ui-list__item', className)}>
+  <li className={classNames('c-ui-list__item', className)}>
     { children }
   </li>
 )
