+++ nebula-components/_components.scss
@@ -16,3 +16,4 @@
 @import './components/table';
 @import './components/tabs';
 @import './components/text-input';
+@import './components/ui-list';
