+++ nebula-components/components/_ui-list.scss
@@ -1,0 +1,33 @@
+@mixin nb-ui-list-spacing() {
+  $class: 'ui-list';
+  $modifier: '--spaced';
+  @each $key, $value in $nb-list-spacing {
+    $compound-class: '#{$nb-namespace}c-#{$class}#{$modifier}-#{$key}';
+
+    .#{$compound-class} > .#{$nb-namespace}c-#{$class}__item {
+      padding-top: $value;
+      padding-bottom: $value;
+    }
+
+    @each $bp-key, $bp-value in $nb-breakpoints {
+      @include nb-respond-to($bp-key) {
+        .#{$compound-class}#{$nb-breakpoint-class}#{$bp-key} > .#{$nb-namespace}c-#{$class}__item {
+          padding-top: $value;
+          padding-bottom: $value;
+        }
+      }
+    }
+  }
+}
+
+@if $nb-use-ui-list {
+  .#{$nb-namespace}c-ui-list {
+    @include nb-bare-list;
+  }
+
+  .#{$nb-namespace}c-ui-list__item + .#{$nb-namespace}c-ui-list__item {
+    border-top: 1px solid $nb-color-base-border;
+  }
+
+  @include nb-ui-list-spacing;
+}
