--- nebula-components/objects/_ui-list.scss
@@ -1,11 +1,0 @@
-@if $nb-use-ui-list {
-  .#{$nb-namespace}o-ui-list {
-    @include nb-bare-list;
-  }
-
-  .#{$nb-namespace}o-ui-list__item + .#{$nb-namespace}o-ui-list__item {
-    border-top: 1px solid $nb-color-base-border;
-  }
-
-  @include nb-list-spacing('ui-list', 'bottom');
-}
